+++ client/src/entities/timer/ui/format-time.ts
