+++ client/src/entities/timer/ui/index.ts
@@ -1,2 +1,2 @@
-export * from "./formatTime";
+export * from "./format-time";
 export * from "./timer-view";
